#80 move some server castling code around, better organization. add more consts for piece color, type
