+++ src/backend/services/chess/chessConsts.py
@@ -1,3 +1,4 @@
+from enum import Enum
 
 wkrStartTile = "77"
 wknStartTile = "76"
@@ -16,3 +17,23 @@ bqStartTile = "03"
 bqbStartTile = "02"
 bqnStartTile = "01"
 bqrStartTile = "00"
+
+ROOK = "Rook"
+KING = "King"
+QUEEN = "Queen"
+BISHOP = "Bishop"
+KNIGHT = "Knight"
+PAWN = "Pawn"
+
+BLACK = "Black"
+WHITE = "White"
+
+class MoveType(Enum):
+    CASTLE = "Castle"
+    EN_PASSANT = "En Passant"
+    NORMAL = "Normal"
+    # CAPTURE = "Capture"
+
+class MoveSide(Enum):
+    QUEENSIDE = "Queenside"
+    KINGSIDE = "Kingside"
